v1.4.1 Using the repository Listing processors as a test

+++ src/main/java/be/thomasmore/pcbuilder/models/GPU.java
@@ -16,16 +16,15 @@ public class GPU {
     private String interfaceType;
     private String clockSpeed;
     private String color;
-    private Integer price;
+    private Double price;
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
-
     public Integer getId() {
         return id;
     }
